requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
libtorrent==2.0.9
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2